refactor: merged the card-list component tests into one file
