+++ src/_test_/card-list.test.tsx
@@ -2,7 +2,23 @@ import { render, screen } from '@testing-library/react';
 import { describe, expect, it } from 'vitest';
 import { CardList } from '../components/card-list/card-list';
 
-describe('Renders correct number of items when data is provided', () => {
+describe('CardList component', () => {
+  // Shows spinner when loading is true
+  it('Shows loading state while fetching data', () => {
+    render(<CardList animalsList={[]} loading={true} error={null} />);
+    const spinner = screen.getByTestId('spinner');
+    expect(spinner).toBeInTheDocument();
+  });
+
+  // Shows message when no data is available
+  it('Displays "no results" message when data array is empty', () => {
+    render(<CardList animalsList={[]} loading={false} error={null} />);
+
+    const message = screen.getByText(/no animals found/i);
+    expect(message).toBeInTheDocument();
+  });
+
+  // Renders correct number of cards
   it('renders correct number of items when data is provided', () => {
     const mockAnimals = [
       {
